Fixed metal reflections getting noisier over time.

+++ source/demo-dll/shaders/raytracing/pathtrace-integrate.hlsl
@@ -24,7 +24,7 @@ void cs_main(uint3 dispatchThreadId : SV_DispatchThreadID)
         RWTexture2D<float3> historyBuffer = ResourceDescriptorHeap[g_constants.historyBufferUavIndex];
         float3 historyColor = historyBuffer[dispatchThreadId.xy];
 
-        float3 avgColor = ((g_constants.historyFrameCount - 1.f) * historyColor + currentColor) / g_constants.historyFrameCount;
+        float3 avgColor = ((g_constants.historyFrameCount - 1.f) * historyColor + currentColor) / (float) g_constants.historyFrameCount;
         historyBuffer[uint2(dispatchThreadId.x, dispatchThreadId.y)] = avgColor;
     }
 }
